deps: use dependencyGraphBuilder.buildFullDependencyGraph (#748)
@stephaniewang526 @elharo 
CC: @yangnuoyu 

I wanted to see whether we need a release of `com.google.cloud.tools:dependencies`. It does not seem needing a release.

With dependencyGraphBuilder.buildFullDependencyGraph, which is marked public in `com.google.cloud.tools:dependencies:1.4.3`. This "dashboard" module no longer needs the copy of the Java code in "com.google.cloud.tools.opensource.cloudbomdashboard.dependencies" package. (and thus this PR deletes the package)