test(spanner): add Spanner test workflow for express release train
Introduces a new GitHub Actions workflow to perform compatibility testing with the Spanner library analogous to the release train.

The workflow utilizes a new shell script (`tests/spanner/test-spanner.sh`) that downloads the `java-spanner` repository at the version that is currently in the Libraries BOM, updates its `sdk-platform-java-config` dependency version to also match the version in the Libraries BOM, and executes the Spanner build process.

In this commit we are only running the Spanner unit tests, but this can later be expanded to include integration tests.